Rename edge to slot (#1134)
This commit changes the term "edge" to "slot" where it actually refers
to a field of an object or a variable on the stack or in any global data
structures. Notably, the trait `Edge` is renamed to `Slot`, and related
types and methods are renamed, too. We still use the term "edge" to
refer to edges in the object graph, regardless whether the edge is
represented by a slot or by an object reference pointing to the target.

The work packet trait `ProcessEdgesWork` and its implementations still
use the term "edge". Although it was originally designed to process
slots, it can be used (and is currently actually used) in node-enqueuing
tracing as well as weak reference processing, in which case it is used
as a provider of the `trace_object` method which traces object graph
edges represented as `ObjectReference` to the target objects.

Note: This commit only does renaming, and does not change the program
logic. The behavior of the program should be exactly the same after this
change.

Fixes: https://github.com/mmtk/mmtk-core/issues/687