Add SweepChunk to native MarkSweepSpace (#1158)
This PR re-introduce the `SweepChunk` work packet to the native
`MarkSweepSpace` when using eager sweeping. The main intention of this
PS is fixing a pathological case where there is only one mutator and the
single `ReleaseMutator` work packet cannot be parallelized.

The algorithm is unchanged for lazy sweeping.

When doing eager sweeping,

1. We first release all unmarked blocks in `MarkSweepSpace::abandoned`
and each mutator.
2. And then we sweep blocks in parallel using `SweepChunk` work packets.
3. We then go through all "consumed" blocks and move them into
"available" lists if they have any free cells.

In step (1), we release blocks for each mutator in `ReleaseMutator`.
Releasing blocks is very fast, so parallelism is not a bottleneck.
During step (2), all blocks are either unallocated or marked, so we
don't need to move any block out of linked lists, avoiding the data race
we observed in https://github.com/mmtk/mmtk-core/issues/1103. Step (3)
is done by one thread, but it is fast enough not to be a performance
bottleneck.

We also introduced a work packet `ReleaseMarkSweepSpace` which does what
`MarkSweepSpace::release` did, but is executed concurrently with
`ReleaseMutator` work packets. In this way, we don't do too much work in
the `Release` work packet, which is a problem we discussed in
https://github.com/mmtk/mmtk-core/issues/1147.

We removed the constant `ABANDON_BLOCKS_IN_RESET` because it is
obviously necessary to do so. Otherwise one mutator will keep too many
blocks locally, preventing other mutators to get available blocks to
use.

We added an USDT tracepoint in `SweepChunk` in both `MarkSweepSpace` and
`ImmixSpace` so that we can see the number of allocated blocks a
`SweepChunk` work packet visited in the timeline generated by eBPF
tracing tools.

We also changed `immixspace::SweepChunk` so that it now *asserts* that
the chunk is allocated rather than skipping unallocated chunks.
`ChunkMap::generate_tasks` already filters out unallocated chunks.

Fixes: https://github.com/mmtk/mmtk-core/issues/1146