Extensible eBPF timeline attributes (#1162)
The purpose of this PR includes:

1. Providing a systematic way to add additional attributes to events on
the timeline generated from eBPF traces.
2. Adding additional USDT probes in mmtk-core and display more useful
information on the timeline.
3. Allowing users, especially VM binding developers, to extend the
tracing scripts to visualize additional USDT probes in the binding.

In previous PRs https://github.com/mmtk/mmtk-core/pull/1154 and
https://github.com/mmtk/mmtk-core/pull/1158, we use "meta" events to add
attributes to work packets, including the number of slots processed by
ProcessEdgesWork and the number of allocated blocks visited by
SweepChunk. This PR refactors the scripts to standardize the use of
"meta" events. Specifically

- The Rust program executes USDT trace points to expose additional
details (such as the number of slots processed) during the execution of
a work packet
-   `capture.bt` emits "meta" events for such USDTs, and
- `visualize.py` adds attributes to the current unfinished GC event
and/or work packet event of the current thread.

In this way, the information about a GC or a work packet can be
gradually added by multiple USDT trace points, providing more
flexibility than having to provide all information at the beginning or
the end of the work packet.

Using this mechanism, we added a few more USDT trace points in
mmtk-core, including:

- Generation plans trigger a USDT trace point when they have decided if
the current GC is a full-heap GC. The information is displayed on the
big "GC" bar (in "Thread 0") for each GC on the timeline.
- Immix-based plans trigger a USDT trace point when they have decided if
the current GC is a defrag GC. Displayed on the "GC" bar, too.
- Root-scanning packets now have "roots" attributes showing how many
root slots or root nodes are reported.
- ScanObjects work packets (as well as ProcessEdgesWork packets that
scan objects immediately) now display the number of objects scanned.

We also introduce extension mechanisms to `capture.py` and
`visualize.py`. Both of them now accept the `-x` or `--extra` command
line argument. Specifically,

- `capture.py` uses `-x` to specify an additional script to be appended
after the `capture.bt` script.
- `visualize.py` uses `-x` to specify a Python script to handle unknown
event names.

The extension mechanism allows VM binding developers to insert custom
USDT trace points in the VM binding, and extend the timeline tools to
visualize those custom events.