Extra assertions for mutators and epilogues (#1182)
This commit adds assertions about the number of mutators and the list of
mutators returned from the VM binding.

Also asserts that epilogues are not silently dropped before they are
executed. It could happen if a buggy VM binding returns a number of
mutators that does not match the actual list of mutators.

This should detect bugs like https://github.com/mmtk/mmtk-ruby/issues/84