Heap traversal (#1174)
This PR adds a heap traversal API `MMTK::enumerate_objects` which
enumerates all objects in the MMTk heap at the time of calling.

We added `SideMetadataSpec::scan_non_zero_values` to support enumerating
objects by scanning the VO bit metadata. It can also be used for
scanning other side metadata when needed. Note, however, that it is
inefficient (but should work correctly) if the metadata is not
contiguous or the metadata has more than one bit per region. If there is
any need for scanning such metadata, we need further refactoring.