Performance history canary
This PR adds a "canary" build to the performance regression CI of
OpenJDK.  The "canary" is a chosen revision of mmtk-core and
mmtk-openjdk that is tested alongside each merged PR.  The performance
of the "canary" should not change unless there is an environment change.
Spotting a change in the "canary" performance can help us identify
environment changes that are unintended or otherwise unnoticed.

TODO: Add a script to select the latest release tag as the canary
instead of hard-coding a version.