Fix a race between forwarding bits and VO bits. (#1214)
The current code sets the forwarding bits before setting the VO bit when
copying an object. If another GC worker is attempting to forward the
same object, it may observe the forwarding bits being `FORWARDED` but
the VO bit is not set. This violates the semantics of VO bits because VO
bits should be set for both from-space and to-space copies. This will
affect VM bindings that assert slots always refer to a valid object when
scanning objects and may update the same slot multiple times for some
reasons.

This revision provides a mechanism to ensure that all necessary metadata
are set before setting forwarding bits to `FORWARDED`. Currently it
affects the VO bits and the mark bits (which are used to update the VO
bits in Immix-based plans). It may be used for other metadata introduced
in the future.