Use modern syntax for optional dependencies (#1229)
Use the "dep:" prefix to specify optional dependencies in Cargo
features.

An optional crate dependency implicitly generates a feature of the same
name, and can be leaked to the user of the current crate. But if a
feature specifies a crate dependency with the "dep:" prefix, it will not
implicitly generate the feature, hiding it from the users. The "dep:"
prefix was introduced in Rust 1.60.