Annotate mmap ranges using PR_SET_VMA (#1236)
We demand that every invocation of `mmap` within mmtk-core to be
accompanied with an "annotation" for the purpose of the mmap. On Linux,
we will use `PR_SET_VMA_ANON_NAME` to set the attribute after `mmap` so
that it can be seen in `/proc/pid/maps`. This will greatly improve the
debugging experience.