Collect live bytes per space, and report by space (#1238)
The current `count_live_bytes_in_gc` feature adds the size of all live
objects and compare with the used pages reported by the plan. There are
two issues with the feature:
1. VM space is not included in the used pages reported by the plan, but
the live objects include objects in the VM space. So the reported
fragmentation/utilization is wrong when the VM space is in use.
2. Spaces/policies have very different fragmentation ratio. Reporting
the fragmentation for the entire heap is not useful.

This PR refactors the current `count_live_bytes_in_gc` feature so we
collect live bytes per space, and report by space.