Allow setting object metadata for VM space objects. Expose VO bit under a feature. (#1248)
This PR changes a few things for vo bit:
1. Add a function to set object metadata for an object in the VM space:
`MMTK::initialize_vm_space_object`.
2. Add a feature `vo_bit_access` to expose VO bit and a binding may use
it at its own risk.
3. Mark VO bit side metadata base address only avilable for 64 bits.

The second is needed for Julia. The Julia binding uses MMTk immortal
allocation or VM space for a region of memory, and pop the regions with
boot image objects with no clear way to identify each object. The easist
workaround is to bulk set VO bit for the region. The problem from this
is that MMTk cannot identify valid objects in those regions. However,
Julia binding only uses VO bit for pinning objects. Objects in the
immortal space or the VM space will not be moved so failing to pinning
objects in those regions is benign. Currently the Julia binding
duplicates a bunch of side metadata code to bulk set VO bit only using
the VO bit side metadata base address. See
https://github.com/mmtk/mmtk-julia/pull/200