Special topic chapter for finalizers and weak references (#1265)
This PR adds a special topic chapter in the Porting Guide for supporting
finalizers and weak references. This topic is frequently asked and
somewhat complex, and needs a dedicated chapter.

We also updated the doc comments of the `Scanning::process_weak_refs`
API to add code example of the intended use case, and warn the users
about potential pitfalls.