Fixing MSRV-breaking dependencies (#1284)
Some dependencies started to require MSRV above our current MSRV.

The build dependency `built` transitively depends on some crates from
the ICU4X project that recently started to depend on Rust 1.81, such as
`litemap`. The dependency on ICU4X is completely unnecessary, and we
removed it from our dependency tree by forcing the use of a particular
version of `idna_adapter`. See: https://docs.rs/crate/idna_adapter/1.2.0

The dev dependency `criterion` depends on `ciborium` which depends on
the `half` crate. Since v2.5.0, `half` started to depend on Rust 1.81.
`ciborium` needs to be fixed because its MSRV is 1.58 and shouldn't
depend on a crate that requires Rust 1.81. We lock the version of the
`half` to 2.4.1. This kind of problem should be properly addressed with
the new MSRV-aware dependency resolver introduced in Rust 1.84. See:
https://doc.rust-lang.org/cargo/reference/resolver.html#rust-version