Aggregate live bytes info in on_gc_finished (#1292)
This PR moves the call to `aggregate_live_bytes_in_last_gc` from the
`Release` work packet to `on_gc_finished`. Spaces may do release work in
parallel with the `Release` work packet, and
`aggregate_live_bytes_in_last_gc` accesses reserved pages of spaces.
This leads to a data race. We observed the reserved pages in the
collected live bytes stats to be larger than the actual reserved pages
at the end of a GC, as some pages were not released yet.