Check if an object is in our heap before using VM map during counting live bytes (#1289)
This PR fixes a bug in `increase_live_bytes`. If an object in the VM
space is scanned, the method will be called with an object reference
into the VM space, which may not be in our heap range. Then we get a
space descriptor from the VM map, and we will see panics in the
following code.

https://github.com/mmtk/mmtk-core/blob/129362d4f39a1191562d276c42f5b03e52aa287a/src/util/heap/layout/map64.rs#L209

https://github.com/mmtk/mmtk-core/blob/129362d4f39a1191562d276c42f5b03e52aa287a/src/util/heap/layout/map32.rs#L258

This PR works around the problem by checking if the object is in our
heap range first.