Allow configuring each Immix space to be non moving (#1305)
We plan to use non-moving Immix as a non-moving space. In this case, we
will have two Immix spaces, one allows moving, and the other does not.
We currently use global constants to control whether Immix moves objects
or not, and we cannot have two Immix spaces that are configured
differently.

This PR removes those constants, and adds a flag `never_move_objects` in
`ImmixSpaceArgs` to indicate whether this Immix space disallows moving.