Make chunk map as global side metadata, and each space only lists its own chunks using chunk map (#1304)
The chunk map records the lowest chunk and the highest chunk used by a
space, and accesses all the chunk states (local side metadata) in the
range. There are multiple issues when we use discontiguous spaces.
1. Chunks in the range may be ~~used~~ unused, thus the local side
metadata may be unmapped.
2. Chunks in the range may belong to other spaces. If other spaces also
use the chunk map, they will list all the chunks even when the chunks do
not belong to the space.

This PR makes the chunk state side metadata as a global side metadata.
We need the chunk map for plans that use the chunk map. As we will use
mark sweep or nonmoving immix as the non moving policy, they both use
the chunk map, and the non moving policy is included in every plan, we
just include the side metadata for every plan.

This PR fixes https://github.com/mmtk/mmtk-core/issues/1197.