Add MultiLevelMapper
This is a replacement for `FragmentedMapper`.  It does not use hash
tables and is lock-free when looking up or lazily initializing the slabs
table.  It is intended to eliminate the pathological case of
`FragmentedMapper` where a hash collision can result in many locking
operations to find the matching or vacant slab table entry.