Replace FragmentedMapper with TwoLevelMmapper (#1337)
This PR replaces `FragmentedMapper` with `TwoLevelMmapper`.

`TwoLevelMmapper` does not use hash tables. It is also lock-free when
looking up `MapState` entries, and also lock-free when lazily
initializing slabs. It is intended to eliminate the pathological case of
`FragmentedMapper` where a hash collision can result in a long-distance
linear scan and many locking operations.

Fixes: https://github.com/mmtk/mmtk-core/issues/1336