Make Mmapper a struct, and existing impls storages
We make Mmapper a struct instead of trait.  It will be solely
responsible for transitioning the map states.

Existing implementations, including `ByteMapMmapper` and
`TwoLevelMmapper`, become the storage backends of `Mmapper`.  They are
only responsible for maintaining the (one-level or two-level) array of
`Atomic<MapState>`, and provide high-level functions for bulk-setting
and bulk-transitioning map states.

Methods of `MapState` are moved to `Mmapper`.