Workaround eBPF errors
The kernel seems to think the map variable `@every` is nullable, even
though we did assign an initial value in the BEGIN block.  We workaround
this problem by using an integer literal (expanded from Python template
variable `$EVERY`).

The USDT `process_root_nodes` may not exist in the compiled .so file of
the VM binding because the methods that contain this USDT are generic
methods that are instantiated on demand.  VMs that do not use
object-enqueuing tracing (such as OpenJDK) may not have this USDT.  We
added the `--no-root-nodes` option to skip that USDT.

Added some debugging options to `capture.py`, including `-d` and `-v`.