Rename Workbucket's active flag to open, add enable flag (#1373)
This PR refactors work bucket to allow disabling a bucket and to support
concurrent GC.
* Add a field `WorkBucket::enabled`: if a bucket is disabled, it behaves
the same as if the bucket/stage does not exist for scheduling, except
that users can add work to a disabled bucket. There are mostly two use
cases for this: 1. a plan can disable buckets that the plan does not
use, and 2. a concurrent plan may push concurrent work to the disabled
concurrent bucket during STW, and enable the bucket at the end of a STW.
* Add `WorkBucketStage::Concurrent`: this is unused in this PR.
* Clean up most checks about stages in `scheduler.rs`, and move them to
`work_bucket.rs`.
* Add a reference to the scheduler in `BasePlan`.

---------

Co-authored-by: Kunshan Wang <wks1986@gmail.com>