Update comment and assert mappable bytes.
Change the comment to say that LOG_MAPPABLE_BYTES is the address space
size the MapStateStorage is able to handle instead of the address space
of the machine.  We also let `Mmapper` expose a method
`log_mappable_bytes` for querying and verification.

We assert when instantiating `MMTK` that `VMLayout::log_address_space`
is not greater than the `log_mappable_bytes` of the chosen `Mmapper`
implementation.