Refactor both Mmapper impls into ChunkStateMmapper (#1369)
The two existing `Mmapper` implementations, namely `ByteMapMmapper` and
`TwoLevelMmapper` have much duplicated code.

We introduce `ChunkStateMmapper` which implements a chunk-grained
`Mmapper` backed by a logical array of `MapState` elements, one for each
chunk. It is responsible for actually calling `mmap` on memory regions.
The `ByteMapMmapper` and `TwoLevelMmapper` become `ByteMapStateStorate`
and `TwoLevelStateStorage`, respectively, and they are only responsible
for holding the per-chunk states and iterating over ranges of memory to
help `ChunkStateMmapper`. We retains the trait `Mmapper` so that we can
still implement other memory mappers that work at different
granularities.

`MapState` is now a pure data type, and state-transitioning methods
previously implemented in `impl MapState` are inlined into methods of
`ChunkStateMmapper`.

We also removed the constant `MMAP_CHUNK_BYTES`. It was defined as
`BYTES_IN_CHUNK`, but it seemed to convey the message that the "mmap
chunk" is special and can have a size that is different from "chunk". We
instead added a function `Mmapper::granularity()` which returns the
actual granularity of the mmapper.