Remove unused methods and types related to memory mapping (#1383)
We remove the `protect` method and the `Mmapper::Protected` state. This
is because Mmapper cannot protect the memory at the granularity desired
by its users. Spaces that wish to protect pages of memory should
directly call `mprotect` while the space still owns the memory.

We also remove the unused `eagerly_mmap_all_spaces` method. It has never
been implemented or used since it was added to JikesRVM MMTk.

We also remove `Space::ensure_mapped`. It was used for the boot image
back in JikesRVM. We now use the `VMSpace` for boot images, and it can
call `MMAPPER.mark_as_mapped` to mark the memory as mapped in the
`Mmapper` without actually calling `mmap`, assuming the VM has already
mapped the memory. So we remove the redundant `Space::ensure_mapped`
method.

Fixes: https://github.com/mmtk/mmtk-core/issues/1348