Concurrent Immix (#1355)
We add a concurrent Immix plan. It can do concurrent marking
(non-moving), and falls back to stop-the-world Immix collection with
opportunistic defragmentation.

We add a snap-at-the-beginning barrier to support concurrent Immix.

Now the plans control when to clear side unlog bits. This allows the
same policy (specifically the `ImmixSpace`) to be used by different
plans and clear the unlog bits at different time.

---------

Co-authored-by: Yi Lin <qinsoon@gmail.com>
Co-authored-by: Kunshan Wang <wks1986@gmail.com>