Fix LargeObjectSpace::initialize_object_metadata (#1393)
It did not set the VO bit when allocating as live (by ConcurrentImmix).
We fixed that. It's always set as long as VO bit is enabled.

We simplified the logic for setting the mark/nursery bits. We no longer
load from the bits.

We always assert the global unlog bit is unset as long as the LOS needs
unlog bits.

We also remove the `alloc` argument of
`SFT::initialize_object_metadata`. It came from the `initializeHeader`
method back in JikesRVM, and was used to distinguish between allocations
by mutators and copying allocations by GC threads. In the current Rust
MMTk, we depend on `PolicyCopyContext::post_copy` to set metadata after
copying, and `SFT::initialize_object_metadata` is only called after
allocation. So the `alloc` argument is now dead code and can be removed.