Update doc comment for Slot
We make it clear that a `Slot` points to a slot instead of being the
slot itself.  We also emphasize that `Slot` can be copied, and the
copied `Slot` points to the same slot.

We also reorganize the doc comment into sections.

We also replace overly general wording with more specific examples, such
as explicitly mentioning JVM, CRuby and JS engines.