Update doc comment for Slot (#1409)
We make it clear that a `Slot` points to a slot instead of being the
slot itself. We also emphasize that `Slot` can be copied, and the copied
`Slot` points to the same slot.

We also reorganize the doc comment and explain the relation between
slots in general and the `Slot` trait in MMTk.

We also use `SimpleSlot` and `OffsetSlot` as examples of what an
implementation of `Slot` can contain.