Fix LOS object enumeration and refactor TreadMill
We fix a bug when enumerating objects in LOS.  During Prepare of
ConcurrentImmix, we flip the from and to space of the TreadMill, and
then enumerate objects in the LOS to set their unmark bits.  However,
the object enumeration code only enumerates objects in the to-space.
Because we have flipped all objects to the from-space, object
enumeration visited no objects at all.  Consequently, no objects in LOS
have unlog bits during concurrent marking, and the SATB barriers will
not be applied to any objects in the LOS.

Now `TreadMill::enumerate_objects` take an additional argument `gc`.
When it is set to true, we enumerate objects in both from-spaces and
to-spaces.  We do this when setting and clearing unlog bits, and fix the
bug.

We also refactored the `TreadMill` so that all of its `HashSet` fields
are protected by one single `Mutex`, because it is not safe to enumerate
one `HashSet` while another thread modifies another.

We also renamed the hash sets to `nursery.from_space`,
`nursery.to_space`, `mature.from_space` and `mature.to_space`,
respectively, for consistency.