Let CI use both OpenJDK 11 and OpenJDK 21 (#1419)
When running tests using OpenJDK, including binding tests and
performance tests, we use both the `jdk-11` branch and the `jdk-21`
branch of the repo https://github.com/mmtk/mmtk-openjdk

Specifically, for binding tests, including both minimal tests and
extended tests, we split OpenJDK 11 and OpenJDK 21 so that they are
treated as two different bindings, and the CI will run tests for OpenJDK
11 and OpenJDK 21 in parallel.

For performance tests, we simply use the `jdk-11` branch instead of the
`master` branch. Since it points to the same branch as the current
`master`, its behavior should not change.